refactor: strategy 패키지 삭제
